requirements.txt
discord.py
google-api-python-client
google-auth
google-genai
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
sentence-transformers
torch
llama-cpp-python
python-dotenv
faiss-cpu
ragas
datasets
numpy
transformers
sudachipy
sudachidict_core
rank-bm25
jsonref
